SCRUM-36 restructure
