--- src/assets/components/GameLobby/GameButtons/index.jsx
@@ -1,3 +1,0 @@
-import GameButtons from "./GameButtons";
-
-export default GameButtons;
