--- src/assets/components/GameLobby/GameInfo/index.jsx
@@ -1,3 +1,0 @@
-import GameInfo from './GameInfo';
-
-export default GameInfo;
