--- src/assets/components/GameLobby/PlayerList/index.jsx
@@ -1,3 +1,0 @@
-import PlayerList from './PlayerList';
-
-export default PlayerList;
