--- src/assets/components/GameLobby/index.jsx
@@ -1,3 +1,0 @@
-import GameLobby from "./GameLobby";
-
-export default GameLobby;
