--- src/assets/containers/GameLobbyContainer/index.jsx
@@ -1,2 +1,0 @@
-import GameLobbyContainer from "./GameLobbyContainer";
-export default GameLobbyContainer;
