+++ src/containers/GameLobbyContainer/components/GameInfo/GameInfo.jsx
